Add package-lock.json for dependency tracking
A package-lock.json file has been added to the project. This file provides version information about the current state of the project dependencies. This helps ensure the project functions properly even when new versions of dependencies are released, by locking in the versions that are known to work with the project.